Added ConstructorInterceptor
Inorder to enable intercepting constructors such as XMLHttpRequest in
the window, we need to be able to intercept constructors in general.
Constructors are just like functions, the only difference is that their
original function should be called with a 'new' operator rather than
using .call or .apply on the function object.

This commit adds the feature along with a test. The types in the
FunctionInterceptor is also enhanced to support better type inference on
the constructor's arguments and return type.