Generic interception for any function
Added a simple mechanism that any function can be intercepted
generically. The new mechanism attaches the interceptor to the function
itself so that it can passed around.

Also fixed a bug where function properties were not correctly copied if
the original function was passed to the constructor of the
FunctionInterceptorBase.

This is going to be useful for intercepting event handler callbacks.