Move wrap/unwrap to FlowletManager to enable external use.
Although we capture all of the browser based api and wrap them properly,
there might be cases that the application itself wants to wrap certain
callbacks to be flowlet safe.
To support this, moved the `wrap` and `unwrap` methods to FlowletManager
so that they can be called by application, or even extended by
overriding the class.