Removed *.js files from the repo.
We will ignore all the .js files in the git. Originally I kept them
there in case anyone wanted to just import the repo and work with it.
But it is a pain to keep track of.
If/when we wanted to post these packages to npm, we can choose to
include the .js files then.
And anyone who pulls in the code, can easily build it now.