Add flowlet constructor to FlowletManager
Since FlowletManager and Flowlet class be extended, it is easier to have
FlowletManager contain all the information needed for managing flow of
flowlets.
This change allows FlowletManager to also carry the corresponding
constructor of the Flowlet. That way, just passing an instance of the
FlowletManger is enough to create and manage flowlets.