Enhance the push logic with `forkReason`.
FlowletManager.push accepts an optional second parameter for pushing new
flowlet. The base class does not do anything with the `forkReason` but
the typical usage of that model is forking the flowlet if there is a
reason specified.
This logic is now added to the base class to eliminate the need for
sub-classing the FlowletManager in most cases.

Also, made the .top() method faster by cacheing the last value and
avoiding the array lookup.