Basic React Element visitor
In order to make it easier to process various types of React components
and elements, this visitor facility basically hides the internals of
React and simply calls back visitor functions depending on the type of
node/element that is passed to it.

Note that to test this class, we need a whole React project setup. There
is a separate package added later that does this outside of thie
package.