Create the hyperionAutoLogging target
To ensure we can build a clean library with no dependencies on external
libraries (e.g. react), I had to change the all imports of React to type
imports. This commit creates a new "hyperionAutoLogging" target which
exposes "Surface" module in one file. This ensure we can continue to
build the whole project as one clean and react version indepenent file.