Added a Channel to support multi-hooks for callbacks.
If we have too many events to fire, it is not efficient to always create
emepty Hooks for them.
The Channel class allows a typed mechanism to lazy create the Hookx only
when there is a callback registered for it.

This mechanism will allow one module to expose multipe events at once.

Not that this is slightly slower than having dedicated hooks since it
involves a lookup and a check.