Add config to allow selective feature disabling
In the main code base we might use various types of gatekeepers to
control which features are enebled.
Added config flags that can be controlled from main code for these
features.

Also added some simple type utility to allow better uniform handling of
options.

Tested the combinations of flags in the
`hyperion-react-testapp/src/IReact.ts` for various compbinations of
`disableReactFlowlet` and `disableReactPropsExtension` to make sure
everything is working ok.