Remove type parameters from as many init functions as possible.
Since TS is more capable in terms of checking sub-types, we no longer
need to have all types passed type parameters to functions (this was
needed in Flowjs).

By removing the type parameters, we can simply enforce the type
requirements with some interfaces. This means, we can better modularize
the code and don't need to wrap everything in init functions.

This commit specially set the stage for fixing the portal issue next.