Added TimedTrigger utility
This utility is needed by many features in the hyperion-autologging.

It provides a convenient mechanism to run a function only once, either
based on time passed, or explicitly. it is usually a good way of running
a code based on another event, or having a timeout mechanism in it.
It also supports cancelling or delaying the execution.

(Note: this used to be called TimedOnceFunc before in the main repo)