Fixed the link between request/response events in fetch
Since fetch handling need several async processes to read the response
value, we often lose the previous value of `requestEvent` since new
versions override it.
Xhr does not have this problem because the value is one per instance of
XHR.

To achieve the same effect with fetch, copied the value of
`requestEvent` to the Promise instance itself. Since everytime fetch is
called a new Promise instance is returned, we won't lose the
`requestEvent` value in between multiple calls.