Fix multiple callbacks blocking api call with OnArgsObserver
In the original design of `FunctionInterceptor.OnArgsObserver`, any
callback can return `true` to block execution of the original function.
However, in practice, only the return value of the single callback would
work and when having multiple callbacks, the values would be ignored
(default implementation of the Hook).

This commit fixes that by adding a custom multi-callback handler. Added
also a test to ensure this.