Add alflowlet to the normal flowlets & network events for better tracking
As we further harden the flowlet tracking, we might lose the link
between a user interaction event and other events such as network
events.
This commit adds a new internal ALFlowletManager that we separately
track the flowlets created by AL itself. Then we attach them to the
normal flowlets of each process to signal to them what "overlapped" with
their flow.
In this way, we deterministically see the link between events and ui
events.

Fixed a bunch of tests around flowlets as well.
Updated the DynamicSvgComponent in test-app to mount a component after
network response comes in to see the effect of correct flowlet
assignment.