Added a mechanism for persisting data in a session.
Sometimes an application may include multiple pages on which AutoLogging
might be enabled. In that case it needs a single a single session id that links all pages of
the same origin together. This id must survive across page loads.
Same thing applies to EventIndex values and they should be continuous
during on full session.

Using browser's `sessionStorage`, this commit creates a simple mechanism
to have `SessionPersistentData` and used it to handle session id and
event index.