Ensure only plane functions are used for Function Components
There was a case in Meta code that a "special typeof" class component
without a 'render' function was created and passed around. T153020999
The existing code does not detect that and will assume it is a Function
Component. As a result, we might incorrectly call hooks for such
components, which causes React to throw.

The change in this commit will ensure that only "plain functions" are
considered for functional component interception.