Rename onArgs* and onValue* to onBeforeCallArgs* and onAfterReturnValue*
I got several feedback that the original names are not very descriptive
and would prefer something that more explicitly indicate the callbacks
will before function call or after return value is ready.
Hence the rename here.
This is a breaing change, but since we still don't have an official
versioning and releasing process, we'd better make this change now!