Expose options to update element text externally
In some cases, the application might want to process the extracted text from the DOM before it is passed along with the event.
Exposed a new option to have a callback that recieves the source elements as they are being computed.
The mechanism supports both single node and multi node cases.

In the test app, showed how one can further extend the text data for inclusion in the event.