Share repeated logic between capture/bubble phase of ui event
To make sure that our logic stays consistent between capture and bubble
phase of ui event handling, moved the repeated code to the a shared
function.
To make this simpler and more efficient, renamed `captureTimestamp` and
`bubbleTimestamp` to just `eventTimestamp` to be consistent with every
other type of event.

Also changed the event handler loop to handle repeated events in a give
event config.