Add default .metadata to all events
Often application code might want to extend the standard AL events and
add data to be logged.

Although a recomended way would be to extend the event types themselves,
but in some simple cases, it is easier to just add the new data to a
default `.metadata` object that is later logged along with the rest of
data.

This commit adds this feature, along with some tests.
Specially for ui events, the metadata of the capture and bubble events
is eventually merged into one and passed to the actual `al_ui_event`.

Later we will properly add supporting metadata for surfaces as well.