Add metadata support to Surface
This commit enables passing a  metadata object to the Surface component
that then will be passed to the corresponding evants.

Since we emit both `al_surface_{un}mound` and `al_surface_mutation`
events, we should make sure possible changes to the metada from
subscribers of the former is passed to the latter.

Also updates one of the test components to ensure metadata is passed
correctly.