Added onArgsAndValueObserver
In several scenarios, we needed to observe args and then perform some
operation based on some value calculated during args observation.
We often end up with a pattern like the following:
```
let ephemeralValue
fi.onArgsObserverAdd((...)=>{
  ephemeralValue = ...
});
fi.onValueObserver(value=> {
  .. = ephemeralValue ...
});
```
Such an approach is ok and works for non-recursive functions. However,
if a recursive method is intercepted, we then have to actually implement
an stack (and hope for the best).

A more reliable approach is to have args oabservers return a value to be
passed to the value observers.

We could come up with some convoluted signature to allow a context
object to be passed around. However, since we are going to create a new
object anyways, we might as well just create a function object and close
on whatever value that function needs. That also makes the signature
clean.

Added a test to showcase this as well.