rename onArgsAndValueObserver to onArgsAndValueMapper
Although onArgsObserver is a bit more efficient because it does not create a new args array, but for onArgsAndValueObserver we were anyways creating a new function object on every call. Therefore, the incremental overhead of converting the arguments into array is small.

To make sure all possible usecasese are covered and avoid state explosion in the `FunctionInterceptor`, renamed `onArgsAndValueObserver` to `onArgsAndValueMapper` and updated all call sites accordinlgy.

This would be the most generic, but possibly higher overhead approach for interception.