Enhance TimedTrigger to also use idle callback
In some cases, we want a TimedTrigger action that preferrably happen at browser idle times, and if not, after some deadline.

This change enables TimedTrigger to use either `setTimeout` or `requestIdleCallback`.

Now that jest fake timers work again on the latest nodejs. enabled the fake timers in the jest test as well.