Use a new mechanism for reading surface context in class components
For class components, we cannot use the `useContext` hook to read the
context.
We used to use an internal and unsafe mechanism to read the context
value and get the surface.
Instead in this change, we use the constructor interceptor of
ClassComponent to read the context value.
To be sure the right context is assigned, we use the legacy context
mechanism (using `contextType`) to assign the surface context value to
it.
This is a best effort tosupport legacy components and does not guarantee
the 100% coverage, but usually is good enough.

Also added a new test component to the test react app to show case that
the render is getting the right flowlet value in that component.