Fix react flowlet for functional components and disable for class components
with React18, we recently get a lot of errors which seems to be because
of using the internal value of the Context._currentValue.

This commit fixes the issue for functional components by directly
calling the `useContext` and dissable the logic for class components
until we find a safe way for them as well.