Fix __DEV__ value.
We need to ensure __DEV__ is always defined since a lot of the code
directly uses it. We also should make sure this code runs ok in github
as well as the envirinment into which the code is imported.

Fixed the assignment to only update it to false in the JEST environment.