Cleanup and optimizations for surface handling
In preparation for adding the region, updated the ALSurfacePublisher to
better rely on information passed from `al_surface_(un)mount` event
data. There were a few repeated action such as reading from the map, or
reading attributes. THose are now removed.

Also noticed that we are not capturing the flowlet of the surface at the
time of creation, rather we capture flowlet when React calls
`useLayoutEffect` etc.

For now, keeping all these flowlets in the event's metadata.