Emit ui event even when propagation is stopped.
We use a capture/bubble event handlers to track UI events and then
either when bubble runs, or after some timeout will emit the
`al_ui_event`.

However, any event handler of the target element may call
`event.stopPropagation()` at any point and disable triggering of bubble
phase.

If the click navigates away from the page, then we would not have enough
time to wait for TimedTigger to fire and we emit the final event.

Using event interception, we can catch when event handlers change the
behavior and not wait for TimedTrigger.