Replace region concept with surface capability
instead of adding the notion of a hidden surface (as a region) to the
Surface, we can track a list of capabilities that we want the surface
to have.

This way we can in future extend this to also have surfaces for which we
might track visibility, or the size of the area they occupy on the
screen.