Support reading flowlet from payload in function wrappers
We wrap several callbacks to ensure flowlets are correctly created for
each callsite.

For triggerFlowlet, we sometimes was to read it from one of the
arguments of the function (e.g. Event, Promise, ...). These are the
cases that we know actually can be a trigger for the next set of
actions.

This commit is the Step2 of supporting trigger flowlet and only adds
support read reading the value from payload. However, we still don't
have the values created or assigned, hence the changes are still no-op.