Remove or disable inactive code.
There was a left over duplicate code for `useState` that is removed now.

Also, we don't yet support accurate (surface,flowlet) mapping, need to
ensure ALL trigger flowlets are assigned to the root flowlets.
If we assign any value to the surface flowlets, they may mask the actual
trigger inherited from the top.