Add a root for FlowletManager
In many case we would want the flowlets to start from known root so that we can share data among all of them.

This commit adds a FlowletManager.root wich is also used to ensure `.top()` never returns null.
That also simplifies some of the logic of dealing with the top of stack. This helps with assigning trigger flowlet as well.