Fixed access to undefined.
If we have the same function passed to both `setTimeout` and
`addEventListerner` we will endup intercepting and wrapping the same
function once. So, the callback handler will get wrapped once with a
getTriggerFlowlet that only works for one situation.

The current logic might work in most cases where the flowlet inherits
the parents trigger flowlet; but would need to consider if there is a
better way to handle both cases combined.