Rename flowlets to always include ID in the name.
In many cases, we need to distinguish between instances of the flowlets. So, this commit by default includes id in the name.

lso, explicitly renamed the root flowlet to /. This way, the flowlet will look like a URI and is easier to copy/paste in other tools.

Consumers of the flowlet can also add a schema before the name before logging to embed the meaning the name, e.g.
trigger://:1/mounse:2/click:3/.