Split Trigger Flowlet logic into smaller sections
Since we have a root flowlet for all flows and trigger flowlet of the ui
events is assigned to that root, we can in some cases live without some
of the more costly features that pass the trigger flowlet around.

In order to measure the performance and accuracy effects of each
feature, split the logic to be controlled by separate flags. In most
cases when there is not overlapping user interactions, the trigger
flowlet will flow correctly.

Also renamed all of the `disable*` flags to be `enable*` so that they
are more logical with optional values. This make is easier to reason
about the overall logic.

Finally, in the to top AutoLogging module, look at all the dependencies
of the features on react interception to ensure we only activate what is
needed.