Further split the cases for trigger flowlet
We know from measurement that react_callbasks case was adding more perf
overhead, so now breaking it down to individual api to see which one is
the root of the issue.

Also did the same for the react state, although we don't see perf issue
there, but doing the change now to prepare for possible need in the
future.