Add an option to collect default values under a surface
In some cases, the user may not click on any of the interactable elements and application will use the default values in such elements.
This commit, adds an option to collect the value of some of these elements (radio/checkbox inputs, and selectors) and emit them as separte events.
The struture of these events is identical to ui events, as if user had
clicked on the element and chosen the value.

I also updated the test react app to ensure this is working.