Fix element reference for surface mount/unmount events
I realized that we sometimes could have more than one element on the page with the same surface attribute, this can for example happen awhen we have as surface proxy. In that case, any search for the right element could return the wrong value.

Instead, this commit changes tha algorith to always pass a ref an ask react to identify the correct element. Note that, we don't need to merge input ref with our local ref, because when we have a input ref, we avoid adding the wrapper.

This new code generates accurate root elements for the surface events.