Simplify surface mutation publisher
I noticed in ALSurfaceMutationPublisher we store duplicate data in the surface info, and may even update the element without updating the corresponding autologging id.

This commit simplifies the logic by reusing the same object (instead creating two objects) both in the map, and for the event.

This information can be later used by others if they want to access mounting info.