The visibility event to handle Surfaceproxy and also hiding of surface
When we have SurfaceProxy (e.g. React.createPortal) the root of the
Surface will be a different element than the original surface. The added
logic handle those cases as well.

Also updated the logic to track transition to/from visible and fire when
surface is hidden.

For testing, loaded the test app, tried the network tab to load
something, when surface mounted the visibility event fired.
Then resized the window so that the surface hides or become visible
again and check the event firing all the time.