DOM Snapshot publisher
Added logic to add the snapshot of the DOM node that user clicked on.
This can be used by tools to have a visual representation of the UI
element clicked on.

Also changed the signature of `emitALCustomEvent` to allow passing
`relatedEventIndex` and link a custom event to another event.
For now, both emit the snapshot as a custom event and keep it on the
metadata of the original event. But for actual logging, this snapshot
should be dropped from the event.
We might want to add `relatedEvent` also to the payload so lookup
becomes easier later.