Enforce tryParentInteractable text extraction limits to elements with installed handlers/data-*able attributes
In some cases clicking an `input` or "interactive"-type tagged element,  there may be no available text to extract as we start extraction from the targetElement itself.

Mehrdad previously added some functionality to attempt parent element if text extraction returns nothing from the targetElement.

The current implementation enforces that interactableElementsOnly setting is set to false,  to prevent walking up the dom too far.

However there are cases we are experiencing where we don't extract text when we should because of the current guard.