Added a new layout config and enabled filteration in the graph
Added fcose layout algorithm which is already imported in WWW
Added a filteration mechanism to the graph to allow very detailed
control over what is added to the graph. This is just a raw form for
whater is not properly supported in the graph.

Later, will add some more advanced filteration to avoid even calling
methods that add nodes to graph.