Fixed a bug in options not working
Since we create a ref for graph and its event handlers, we should make
sure that we don't close on the initial value of the options to make
sure we can continue to pick it up.
Also, had to ensure the reference to the options object changes to force
react to re-trigger a bunch of logic.