use a flow-session-id instead of domain-session-id
assigning a session id to the top domain means that all tabs and apps on
the same domain will get the same value.
Instead, trying another approach in which a short lived session cookie
is set on the domain and then immediately picked up by the next
application that is openned.
In this way, we can stich together flows between applications, and can
continue to rely on ClientSessionID for 'per tab' session
identification.