Simplify the graph view component
Since the graph options take too much space on the screen, re-ordered
the items and added a menu that can show/hide the options components.
This way more space is assigned to the graph itself.
Also, now we can add footer to the graph, which is left empty for now.