Remove the extra flowlet push that was never poped
As part of
https://github.com/facebook/hyperion/commit/44523f0bb5a49e4b969364902e06ffae10ede621
PR, the uiFlowlet was removed but the pop logic was still left behind.
The pop function will be no op of an empty flowlet is passed to it.
As a result, every ui event was pushing a flowlet that stays on the
stack.
In particular, mouseover would keep pushing flowlets that mess up other
cases.

This fix ensure we stick to the principle that all trigger_flowlet must
be attached to the payload of the main trigger, and then interceptors of
that payload (trigger) should do push/pop.