Separated exported files of each package
Instead of listing what each package exports in the main index.ts, I
created one index per package to have all the data local. This is easier
to track what each package exports. This is also more compatible with
the npm convention.
I needed to export the new global flags feature in another diff and
realized it is hard to to manage these.

This change also add a new hyperionGlobal chunk that matches the
hyperion-global packages.