Compbine all data-*able attributes
We used to add one separate attribute for each type of event to the dom
ndoes.
With this change, we add only one attribute, and then the value
determines which events are added.

In the next diffs, we can drop the check for the exact event and just
check if that element is interactable.