Improve stability of ClientSessionID and EventIndex
We noticed that some of the event indexes at the end of a session
overlap with event indexes of the next session.
We have a scheduller for PersistentData class to lazily write back
updated values to storage. This scheduler may miss the final write
specially if page is busy during logging and then is unloaded (reload or
move to another page).

The new code does a few things to solve this:
- The timer is reduced from 100ms to 20ms. This is still better than
  constantly writing the value back (specially for event index)
- The scheduller is updated to handle page unload and turn go into
  immediate write mode.

I also updated the guid logic to generate better random data with lower
overhead (T164417895).

Hopefully all of these changes reduces the chance of event index
overlaps.