Replace pageURI in the events with URL object instead of string
In many cases, we want the elements of the main page. Instead of using
`window.location.href` everytime and parsing it, we now have a URL
object that represents the main page, and the pass that around.

This way, we can avoid over parsing everytime.

We will also need this in one of the plugin.