Add surface data to all events that carry surface
To avoid multiple look ups of surface data for various events, added
surfaceData along with surface in any event that carry them. This way,
we don't have to look them up all the time and the information will be
available everywhere.

Technically, we can remove the `surface` field, or change its type. But
that might be too disruptive of change across the code.