Improve error messages and memory usage.
When detecting a surface used for multiple instances, we can just report
it once rather than flooding the console.

Also, to have the same memory footprint as before, for now not storing
the visiblity events in the surface tree. This should be identical (even
a bit smaller since there is one less surface map) to before.