Ensure visibility events can access mutation events
In the previous logic, we were storing MutationEvents in a map and fire
visiblity events according to those. There was a condition that if a
surface didn't have mutation event, we would not event try to emit
visibility event.

In the previous commit for merging the maps, this condition was dropped.
Adding it back.
This should also address the internal task (T215187690) that has
captured some null access exceptions.