Track element associated with surface
Now that surface capability will allow disabling mutation event, we
won't have any other way of knowing the associated dom element of the
surface.
To address this, added a `Set<Element>` to `ALSurfaceData` to keep track
of the element value.

This also allows us to know that we are adding multiple instances of the
same surface value.