Add event level metadata to surfaces.
There has been many requests to have an automated mechanism to mark
events under a surface with certain metadata. This has also been a
source of confusion in the past.

Added a new props to <Surface> that one can use and define what metadata
should be added for what event from surface to the  event.metadata.

This simple mechanism should eliminate the need for people to add custom
event handlers and manually move data between surface and UI event.