Attach surface and event name to trigger flowlet data
We already include surface and eventName in the name of the trigger
flowlet. We can also ensure this information is part of the
corresponding flowlet.

This way, if a downstream process needs this data, they won't have to
rely on calling `.toString` on the whole chain and then extracting the
value from the string.