Update surface visibility algorithm to support nested surfaces.
Since a surface is not visibile, we have to search for non-surface roots
to apply intersection observer. However, if we have nested surfaces that
don't add any other content, then we have to go deeper in the dom tree
until we find proper roots.

I updated the algorithm accordingly. Also updated the test app to
re-create such situation. It clearly showed that previous algorith was
not working, but the new one handles the situation well.