Update event hash to be more stable
Based on internal discussions, we decided to have a short and stable
event hash the combines an event id, with one stable parameter of the
event (such as `surface` for most events, or `uri.pathname` for network
events).

Such as has will allow us to name similar events in a consistent way
across all stages of logging.