Fixed incorrectly closed on variables for wrapping for flowlets.
When we wrap functions for trigger flowlet, we close on two variables,
{callflowlet, getTriggerFlowlet}. If the same callback is passed
multiple times, we only wrap it the first time and don't recreate
wrappers again. This is similar to useing `useCallback` with wrong
dependencies.

I fixed the issue in this commit by creating new wrappers around the
original function.

This can potentially increase the memory usage, but we sould see if it
significantly improves the traigger flowlet tracking.