Speeding up local builds
The full build is getting slow. TypeScript 7 is in preview but is
significantly faster (uses go).
I added a "build-go" command for now that we can use to run local builds
faster. This is not integrated in anything else, and stays around until
official release of the tsc in future.