Improve performance of Surface rendering
Currently we use a single map to store ALL full surface names mapped to
their SurfaceData.

This map can get really big, specially since we store both the
interactive and non-interactive paths in it.

During surface rendering, we look for 'non-interactive' names in this
map. I repleace that with a per-node map. Since we know the parent node
already (from context), we just need to lookup in the children of that
node, instead of all surfaces.

Some UI events need to convert a surface path back to a SurfaceData. So,
the original global map is now only maintained for those surfaces.
Later, we can try to eliminate that as well if we can find a faster way
to map elements to surface-data.

Testing on Ads Manager, the typical number if entries in the map went
from 1000 down to about 400 items.