Put surface deletion behind a flag
Noticed that in React strict mode, the useEffect called multiple times,
which causes surfaces to be deleted and not put back again.
Anyways, we are not seeing much saving from adding/removing these
surfaces since most of them are used as user navigates around the UI.
For now, put it behind a flag in case we wanted to enable it immediately
in production.

To help debug such situations, introduced an `al_surface_render` flag
that allows us to track how react calls render vs. mount/unmount the
components. It is actually useful to track rerenders at a high level.
Since this could be expensive, put this event also behind a flag.

Also cleaned up a few other unused flags.