Remove FlowletManager from init options
We originally allowed FlowletManager to be passed as an init options.
But, we never saw anyone try to customize it. Also, the class has events
that one can listen to for tracking push/pops.

In order to further simplify the init functions of all modules, this is
first step to remove this field and allow them to use a global one.
If needed, later we can have an init function to update that global one.