Simplify the init of ALSurfaceContext and expose IReact internals
To simplify most init functions and make more of the logic independent
of the init, expose some of the IReact internals so that other module
can direclty use it. This way, as long as modules are initialized in the
right order code continues to work, but then things are much simpler in
terms of exporing stuff from modules and using in other modules.