Separate SurfaceProxy entirely from ALSurface
With this change, the actual Surface component is completely independent
from SurfaceProxy logic and only handles real surface. The SurfaceProxy
mnages the necessary logic and calls new SurfaceImpl component.

In this way, we reduce the reliance on dom api further, so that
ALSurface itself can be easily used in RN